logic_engine_test.py
import os
import time
import logging
import pandas as pd
from datetime import datetime, timedelta, timezone
from google.cloud import bigtable
from google.cloud.bigtable.row_set import RowSet
from google.cloud import pubsub_v1
import json
import pandas_ta as ta
import mplfinance as mpf
import matplotlib
import threading # For background subscriber
import uuid # For generating unique trade IDs
import csv

# --- Configuration ---
PROJECT_ID = os.environ.get("GCP_PROJECT_ID", "telegram-signals-205cc")
SIGNAL_TOPIC_NAME = "trade-signals-micro"
INSTANCE_ID = "cryptotracker-bigtable"
TABLE_ID = "market-data-1m"

# --- Logging ---
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')

# --- Clients ---
logging.info("Attempting to connect to clients...")
try:
    bigtable_client = bigtable.Client(project=PROJECT_ID, admin=True)
    instance = bigtable_client.instance(INSTANCE_ID)
    table = instance.table(TABLE_ID)
    logging.info("Successfully connected to Bigtable.")
except Exception as e:
    logging.error(f"Failed to connect to Bigtable: {e}")
    table = None

try:
    publisher = pubsub_v1.PublisherClient()
    signal_topic_path = publisher.topic_path(PROJECT_ID, SIGNAL_TOPIC_NAME)
    logging.info(f"Successfully connected to Pub/Sub topic '{SIGNAL_TOPIC_NAME}'.")
except Exception as e:
    logging.error(f"Failed to initialize Pub/Sub publisher: {e}")
    publisher = None
logging.info("Client connections attempted.")


logging.info("INFO: Starting MICRO-SCALP Logic Engine...")

while True:
    time.sleep(1) 