requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
ccxt>=2.0.0
scikit-learn>=0.24.0
plotly>=5.3.0
python-binance>=1.0.0
joblib>=1.0.0
python-telegram-bot>=13.0
google-cloud-pubsub>=2.0.0
firebase-admin>=5.0.0
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0
asyncio>=3.4.3
websockets>=9.1
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
black>=21.7b0
flake8>=3.9.0
mypy>=0.910

# TA-Lib-Precompiled==0.4.25 # Removed due to build issues, using pandas-ta

# API clients
lunarcrush==1.0.0  # LunarCrush API client
krakenex==2.1.0  # Kraken API client

# Telegram integration – v21 introduces native asyncio and message editing helpers
python-telegram-bot==21.2
nest-asyncio==1.6.0  # Allows nested asyncio event loops for compatibility
redis==5.0.7 # For Telegram notifier message editing

# Utilities
pytz==2024.1
python-dateutil==2.8.2 
scipy 

# Backtesting Dependencies
pandas-ta>=0.3.14b
plotly>=5.18.0
seaborn>=0.13.0
matplotlib>=3.8.0
scikit-learn>=1.3.0 

# Google Cloud Dependencies
google-cloud-bigquery>=3.18.0 